// app/common/renderer/lib/ai/types.ts
export interface Versions {
  ios: IOS;
  android: Android;
}

export interface State {
  id: string;
  timeStamp: string;
  versions: Versions;
}

export interface Page {
  id: string;
  name: string;
  description: string;
  module: string;
  states: State[];
}